requirements.txt
numpy
matplotlib
opencv-python
pillow
paddlepaddle
paddleocr
scikit-learn
torch
torchvision
torchaudio
ultralytics